remove ui objects from library
ui should be dev only not framework wth am i doing mang?
well.. its partially because this was an artifact from when the
framework was also the game because i thought i could just export the
stuff the game implemented but ehh maybe that was a bad idea in the long
run. because i was starting to implement things for the game's sake and
not the framework's sake. so hoo rah!

Signed-off-by: nowaaru@proton.me <vert@leanbox.>